Update callers of VersionParser::formatVersion()
Tests were moved to BasePackageTest.
