+++ src/Composer/DependencyResolver/Operation/SolverOperation.php
@@ -12,7 +12,6 @@
 
 namespace Composer\DependencyResolver\Operation;
 
-use Composer\Package\Version\VersionParser;
 use Composer\Package\PackageInterface;
 
 /**
@@ -46,6 +45,6 @@ abstract class SolverOperation implements OperationInterface
 
     protected function formatVersion(PackageInterface $package)
     {
-        return VersionParser::formatVersion($package);
+        return $package->getFullPrettyVersion();
     }
 }
