+++ tests/Composer/Test/Package/Version/VersionParserTest.php
@@ -21,52 +21,6 @@ use Composer\Package\PackageInterface;
 
 class VersionParserTest extends \PHPUnit_Framework_TestCase
 {
-    /**
-     * @dataProvider formattedVersions
-     */
-    public function testFormatVersionForDevPackage(PackageInterface $package, $truncate, $expected)
-    {
-        $this->assertSame($expected, VersionParser::formatVersion($package, $truncate));
-    }
-
-    public function formattedVersions()
-    {
-        $data = array(
-            array(
-                'sourceReference' => 'v2.1.0-RC2',
-                'truncate' => true,
-                'expected' => 'PrettyVersion v2.1.0-RC2'
-            ),
-            array(
-                'sourceReference' => 'bbf527a27356414bfa9bf520f018c5cb7af67c77',
-                'truncate' => true,
-                'expected' => 'PrettyVersion bbf527a'
-            ),
-            array(
-                'sourceReference' => 'v1.0.0',
-                'truncate' => false,
-                'expected' => 'PrettyVersion v1.0.0'
-            ),
-            array(
-                'sourceReference' => 'bbf527a27356414bfa9bf520f018c5cb7af67c77',
-                'truncate' => false,
-                'expected' => 'PrettyVersion bbf527a27356414bfa9bf520f018c5cb7af67c77'
-            ),
-        );
-
-        $self = $this;
-        $createPackage = function ($arr) use ($self) {
-            $package = $self->getMockForAbstractClass('\Composer\Package\BasePackage', array(), '', false);
-            $package->expects($self->once())->method('isDev')->will($self->returnValue(true));
-            $package->expects($self->once())->method('getSourceType')->will($self->returnValue('git'));
-            $package->expects($self->once())->method('getPrettyVersion')->will($self->returnValue('PrettyVersion'));
-            $package->expects($self->any())->method('getSourceReference')->will($self->returnValue($arr['sourceReference']));
-
-            return array($package, $arr['truncate'], $arr['expected']);
-        };
-
-        return array_map($createPackage, $data);
-    }
 
     /**
      * @dataProvider numericAliasVersions
